requirements.txt
python-dotenv
django
aiogram
psycopg2-binary
apscheduler
django-ninja
